this'll fix an older issue.
I removed it's parents, and set it to automatically unlock, it's not a great fix, but it'll at least make it unlockable.

This is certainly a bandaid fix, I'll see if I can actually fix it later, but for now this should fix the bulk. 

This hasn't been fixed in 4 years, so fixing it at all, no matter the quality is justified.